Cleaned up tests
